refactor(ff-stream): split dash_inner.rs into focused submodules
